refactor(connector): return to file per connector and move deps to main
