+++ lib/connectors/index.ts
@@ -1,5 +1,5 @@
-export { Connector } from "./connector";
-export { MongoDBConnector, MongoDBOptions } from "./mongodb/connector.ts";
-export { MySQLConnector, MySQLOptions } from "./mysql/connector.ts";
-export { PostgresConnector, PostgresOptions } from "./postgres/connector.ts";
-export { SQLite3Connector, SQLite3Options } from "./sqlite3/connector.ts";
+export { Connector } from "./connector.ts";
+export { MongoDBConnector, MongoDBOptions } from "./mongodb-connector.ts";
+export { MySQLConnector, MySQLOptions } from "./mysql-connector.ts";
+export { PostgresConnector, PostgresOptions } from "./postgres-connector.ts";
+export { SQLite3Connector, SQLite3Options } from "./sqlite3-connector.ts";
