--- lib/connectors/mysql/deps.ts
@@ -1,4 +1,0 @@
-export {
-  Client as MySQLClient,
-  Connection as MySQLConnection,
-} from "https://deno.land/x/mysql/mod.ts";
