+++ lib/connectors/postgres-connector.ts
@@ -1,11 +1,11 @@
-import { PostgresClient } from "./deps.ts";
-import type { Connector, ConnectorOptions } from "../connector.ts";
+import { PostgresClient } from "../../deps.ts";
+import type { Connector, ConnectorOptions } from "./connector.ts";
 import {
   SQLTranslator,
   SupportedSqlDatabaseDialect,
-} from "../../translators/sql-translator.ts";
-import type { QueryDescription } from "../../query-builder.ts";
-import type { Values } from "../../data-types.ts";
+} from "../translators/sql-translator.ts";
+import type { QueryDescription } from "../query-builder.ts";
+import type { Values } from "../data-types.ts";
 
 export interface PostgresOptions extends ConnectorOptions {
   database: string;
