--- lib/connectors/postgres/deps.ts
@@ -1,1 +1,0 @@
-export { Client as PostgresClient } from "https://raw.githubusercontent.com/deno-postgres/deno-postgres/master/mod.ts";
