+++ lib/connectors/sqlite3-connector.ts
@@ -1,11 +1,11 @@
-import { SQLiteClient } from "./deps.ts";
-import type { Connector, ConnectorOptions } from "../connector.ts";
-import type { QueryDescription } from "../../query-builder.ts";
-import type { FieldValue } from "../../data-types.ts";
+import { SQLiteClient } from "../../deps.ts";
+import type { Connector, ConnectorOptions } from "./connector.ts";
+import type { QueryDescription } from "../query-builder.ts";
+import type { FieldValue } from "../data-types.ts";
 import {
   SQLTranslator,
   SupportedSqlDatabaseDialect,
-} from "../../translators/sql-translator.ts";
+} from "../translators/sql-translator.ts";
 
 export interface SQLite3Options extends ConnectorOptions {
   filepath: string;
