--- lib/connectors/sqlite3/deps.ts
@@ -1,1 +1,0 @@
-export { DB as SQLiteClient } from "https://deno.land/x/sqlite@v2.3.1/mod.ts";
